Minor bindings update ("optional" textures: [GLtexture1] is a type for a valid GL texture, [GLtexture0] is a type for either a valid GL texture, or an uninitialized texture; [GLtexture_zero] is for GL texture 0).
Minor update to TGA loading code: run-length encoded
images are now supported; however, the code does not
handle RLE packets spanning rows (this is because
we chose to go by the specification).

Major update to example 3: Wavefront OBJ file loader.
The example can now show three kinds of meshes:
- for vertex-only meshes, we show a solid color mesh
- for meshes that have no materials defined or no texture coordinates,
  but have normals, we show a mesh illuminated by a simple algorithm
- otherwise, we show texture-mapped and illuminated meshes

Note that converting meshes from the form they are stored
in OBJ files to a form most suited to the task of rendering
is a bit complicated.